Fix DataTableScreen searchable so that it works when included form github.
This was the weirdest bug ever. I ultimately decided to just make a separate delegate class and use that to call delegate methods with a different signature than the ProMotion searchable ones. This seems to work fine, but I'm still at a loss as to why I couldn't just override the ProMotion module delegate methods.