Add a remote_image method that can take a block
Previously it was hard to tell when an image had finished loading from
a URL. Having the ability to call remote_image with a block that will
execute when an image finishes loading makes this super easy. Being
able to do this is useful when displaying spinners as they can now be
removed easily at the appropriate time.

The `remote_image` method is separate from the setter method
`remote_image=` to adhere to the convention of setter methods and to
provide a clear interface.